requirements.txt
asgiref==3.9.1
Django==5.2.6
gunicorn==20.1.0
setuptools==80.9.0
sqlparse==0.5.3
tzdata==2025.2
django-environ==0.11.2
dj-database-url==2.1.0
psycopg2-binary; sys_platform == "linux"
whitenoise==6.6.0
Pillow
django-paypal==2.0
django-cloudinary-storage
cloudinary